Move files into src/ and test/ dirs.
