+++ packages/react-aria-modal/.gitignore
@@ -1,0 +1,1 @@
+!rollup.config.js
