+++ packages/react-aria-modal/src/focus-trap-react.ts
