+++ packages/react-aria-modal/src/focus-trap.ts
