+++ packages/react-aria-modal/test-manual/demo-one.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import AriaModal from '../react-aria-modal';
+import AriaModal from '../src/react-aria-modal';
 
 class DemoOne extends React.Component<{}, {modalActive: boolean}> {
   constructor(props: {}) {
